V02：Pro add commd get zeta mac

+++ HAL/inc/power.h
@@ -18,7 +18,7 @@ void BatteryInit(void);
 
 int8_t CheckBattery(void);
 
-uint16_t CheckRecharge(void);
+uint8_t CheckRecharge(void);
 
 int8_t ReadBattery(void);
